completely rewritten lookup
The key change is to have only a single request to the ETCD cluster,
getting all the relevant data. It takes nearly the same time as a
normal request (used before), so we get an upper bound on the request
time. Before, additional requests were required to get defaults data
on different resource records, especially for an ANY request, which
timed out often. Additionally we don't need to use requests with a
revision number, because transactions do that for us implicitly.

Further we don't need to cache defaults any more and thus the advice to
have a separate cluster for DNS is dropped. It's now perfectly fine to
have an active (in terms of revision changing) cluster for pdns-etcd3
without sacrificing performance.