Refactored lookup and rewritten a bit, now working again! And more...
At some point from the last big rewrite to now it was broken, but dirty code changes
masked that :-( Now it's clean again and working. I use this version in semi-production.

I could not break this amount of changes down to smaller commits, so here comes the
detailed list of changes (each would have been worth a single commit):

* defaults entries now with QTYPE, id, both or none (more possibilities)
* reading and handling versioned entries
* caching data in memory to avoid requests to ETCD (not very good implementation yet)
  * new setting 'min-cache-time'
* new abstraction layer for ETCD responses (type getResponse)
* now really supporting unknown QTYPEs as plain string entries

Unfortunately, due to the current implementation the setting 'reversed-names' must be true!
See lookup.go:88 for details.