rewritten to using watcher into ETCD; changed data structure; more...
A. High-level description of changes

1. The main thread now sets up two channels for requests from PowerDNS and data changes from ETCD.
It handles those requests sequentially, so no locking is needed. By
using a watcher into ETCD data changes are applied "instantly" with
reloading only the relevant zones (and sub-zones). The only exception is
deleting a SOA record - then the parent zone (or root for a top-level zone) is reloaded.

2. The data structure changed, but only the separators and the
requirement of reversed names. This is done to support (graphical) tools
(like ETCDkeeper) to display entries directory-like, which eases
browsing a lot. Reversed names are now 'baked-in' (required), b/c
reloading a zone incurs a prefix request into ETCD for the zone, which
returns all entries relevant to it in only one call / with only one key,
which improves performance (instead of parsing all(!) of the data again
and again on minor changes). But it is also a requirement for the
browsing feature.

3. Now (2 years later) PowerDNS 4 is the default, PowerDNS v3 is not
supported any more. So the default version changes to 4. Also the
backend protocol problem is described more thoroughly.

4. Made "-options-" a first-class citizen (in the description, but no options implemented yet).

B. Implementation stuff / changes

* go 1.18 now required due to generics (util.go: clearMap)
* dropped caching mechanism, implemented watcher(s)
   * reading in go routines, posting to channels, main thread selecting on them
* added more logging
   * added more details to existing log messages
* dropped parameters 'reversed-names' and 'min-cache-time'
* changed program version syntax (and described it better)
* fixed error handling on missing values (e.g. a missing "port" for SRV became 0 instead of raising an error)
* fixed data version parsing (error'ed when using only major version)
* fixed domain qualified name generation (dataNode.qname)
* more structure in main func, refactored into separate functions
* dropped old code, now reading and parsing data is not duplicated
* reworked name class due to reversed-names requirement, now easier, more straight-forward
* fixed response for NXDOMAIN, an empty array was not the correct answer on that, but a "false" (weird PowerDNS behaviour)