bumped etcd version to 3.5.15
Big jump, but still in go 1.21 line. Had to adapt code a tiny bit. First release with go.mod, it was the easiest to catch up with. The other versions (3.1 - 3.4) all had very complicated dependency problems, would have need to pin many libraries manually.

Since pe3 is only using the client, the changes were only in loading the configuration file. But need to write an automated test for it yet.